Add strEQs, strNEs, for literal final param
If we are comparing two strings, we can skip the actual compare if the
lengths aren't the same (but calculating the lengths ahead of time
negates any savings).  However, it turns out that there are a number of
cases in core where the length of one string is known, and the other
string is a literal, hence its length can be calculated by the compiler.
This macro allows those cases to be easily handled, and with shorter
code.

This eliminates programmers from having to count characters.