APItest/t/utf8_setup.pl: Split function into two
This function outputs a byte string as hex bytes.  A future commit will
want that output without surrounding quotes, so create a version that
doesn't have them.

This also corrects the number of bytes needed to discern that the
overflow happens on 64-bit platforms from 2 to 3.  This error would
be exposed by tests added in future commits.