bytes_from_utf8() Can must do memcpy if all invariant
This function does two passes over the input.  In the first it decides
if the string can be downgraded, and computes the size needed for the
downgraded string.  In the 2nd pass, it does the conversion.

Likely, this wouldn't be called unless there is something to downgrade,
but adding a single 'if' can cause the downgrading to be entirely
skipped, replaced by a memcpy().