Fatalize bit ops on wide characters
These deprecations are scheduled to become fatal in 5.28.

It turns out that a bunch of the code dealing with UTF-8 operands could
be cut out.  The operands are simply converted to non-UTF-8, croaking if
that fails.  Then the operations is done on non UTF-8, using the
existing code.  To preserve current behavior, the result is left not
UTF-8 if the op is "~", and is changed back to UTF-8 for the other ops.

The part of this that took the most time was going through t/op/bop.t,
looking at all the failing tests and trying to determine if the tests
are simply testing the wide character behavior, and hence can just be
deleted; or if they are testing general UTF-8 behavior, and hence should
be modified to not use above-FF code points, while retaining UTF-8ness.
I went through the blame log for each such test, reading the commit
messages.  Most turned out, unless I erred, to be testing wide character
behavior, which is now illegal.