APItest/t/utf8_warn_base.pl: Display mnemonics on error
Part of the testing for this is that the returned flags for problematic
conditions are correct.  This commit adds a routine that will convert
numeric values of the flags into a mnemonic string like FOO|BAR|BAZ.
This makes debugging easier.  The names are not computed unless there is
an error.