APItest/t/utf8_warn_base.pl: Revamp loop to/from utf8
This test file had gotten kinda messy as new tasks were shoe horned into
it.  This cleans it up, and positions it to be easier maintain going
forward.  I tried to minimize the number of changes shown per commit,
but this is the minimal I could get, and since it is a revamp, there are
lots of differences.

Some combinatorial explosion has been removed.

A new subroutine is created which compares the expected vs actually
gotten warnings, and is called in two places, removing duplicated code.

This exposed a bug in very large, hence rare, code points.  It will be
fixed in the next commit.  It was far easier to just make all similar
tests TODO here, removing that in the next commit.  This means this
commit has many passing TODOs