t/op/chop.t: Don't use too large code points
The bug this was testing for requires a code point that will no longer
be legal on 32-bit machines.  So skip there, and revise to use chr() in
the skipped code instead of "\x{}".  The latter would compile even if
execution gets skipped, so would cause it to die.  This also tests the
very highest legal code point on 64-bit machines, which is now illegal,
so test the new very highest one.