APItest/t/utf8_malformed.t: move test to utf8_warn_base.pl
Actually, this test was already in utf8_warn_base, but was executed only
on 64 bit platforms.  It is reasonable to make sure it works on 32 bit
ones, as it is an edge case there as well, in the sense that it is the
first 13 byte code point.

This is the first of a series of commits to remove all the tests in
utf8_malformed, so the entire file can be removed.

utf8_warn_base has been heavily cleaned up, and now has better
infrastructure for more completely testing thant utf8_malformed.  The
two files have much the same logic, and rather than trying to maintain
two versions, it's better to combine them.