I18N-Langinfo/t/Langinfo.t: Reinstate tests, modernize
The comments say that a bunch of tests were disabled due to the poor
support of nl_langinfo() at the time, 2001.  I haven't seen these issues
lately, and so am reinstating many of the tests, using the modern
Test::More that the rest of the file in the meantime had been updated to
use.  The tests that weren't reinstated were because results for them
may vary by platform.